pyproject.toml
[tool.poetry]
name = "drf-template"
version = "0.1.0"
description = "A comprehensive starting point for a new API with DRF"
authors = ["Sean Njela <seannjela@yahoo.com>"]
license = "MIT"
readme = "README.md"

packages = [
    { include = "scripts" }
]

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.2"
django-environ = "^0.12.0"
django-cors-headers = "^4.5.0"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9"
whitenoise = "^6.7.0"
gunicorn = "^23.0.0"
django-rest-knox = "^5.0.2"
redis = "^6.0.0"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
django-celery-results = "^2.5.1"
sentry-sdk = {extras = ["django"], version = "^2.17.0"}
django-redis = "^6.0.0"
drf-spectacular = "^0.28.0"
faker = "^37.1.0"
django-seed = "^0.3.1"
django-extensions = "^4.1"
django-filter = "^25.1"
python-json-logger = "^3.3.0"
requests = "^2.32.4"
rich = "^14.0.0"
mkdocstrings = "^0.29.1"
pymdown-extensions = "^10.16"
mkdocs-mermaid2-plugin = "^1.2.1"
mike = "^2.1.3"
mkdocs-minify-plugin = "^0.8.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^6.0.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
ipykernel = "^6.29.5"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.26"
black = "^24.4.2"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.10.0"

[tool.poetry.scripts]
server = "scripts.django:runserver"
makemigrations = "scripts.django:makemigrations"
migrate = "scripts.django:migrate"
worker = "scripts.celery:run_worker"
beat = "scripts.celery:run_beat"
bump = "scripts.bump:bum_version"
create_dev_env = "scripts.django:create_dev_env"
seed = "scripts.django:seed"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 95
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 95

[tool.flake8]
max-line-length = 95
exclude = [".venv", "migrations", "__pycache__"]